requirements.txt
azure-identity==1.13.0
python-dotenv==1.0.0
pyodbc==4.0.39
SQLAlchemy==2.0.15
